+++ projects/project_0/Project0.StoreApplication.Domain/Models/Customer.cs
@@ -6,7 +6,7 @@ namespace Project0.StoreApplication.Domain.Models
 
   public class Customer
   {
-    public byte CustomerID { get; set; }
+    public int CustomerID { get; set; }
     public string Name { get; set; }
 
     public List<Order> orders { get; set; }
