+++ projects/project_0/Project0.StoreApplication.Domain/Models/Store.cs
@@ -4,7 +4,7 @@ namespace Project0.StoreApplication.Domain.Models
 {
   public class Store
   {
-    public byte StoreID { get; set; }
+    public int StoreID { get; set; }
     public string Name { get; set; }
 
     public string Location { get; set; }
